fix local tests

+++ tests/conftest.py
@@ -64,7 +64,8 @@ def sqlalchemy_connect_url(request, db_url):
 def db_schema(engine, sqlalchemy_keep_db):
     db_exists = sqlalchemy_utils.functions.database_exists(engine.url)
     if db_exists and not sqlalchemy_keep_db:
-        raise RuntimeError("DB exists, remove it before proceeding")
+        sqlalchemy_utils.functions.drop_database(engine.url)
+        db_exists = False
 
     if not db_exists:
         sqlalchemy_utils.functions.create_database(engine.url)
@@ -72,7 +73,8 @@ def db_schema(engine, sqlalchemy_keep_db):
             conn.execute(text("CREATE EXTENSION postgis"))
             conn.commit()
     yield
-    sqlalchemy_utils.functions.drop_database(engine.url)
+    if not sqlalchemy_keep_db:
+        sqlalchemy_utils.functions.drop_database(engine.url)
 
 
 @pytest.fixture
